+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fish: change fuzzy completion prefix to `z!`.
 - Zsh: allow `z` to navigate dirstack via `+n` and `-n`.
+- Fzf: improved preview window.
 
 ## [0.8.2] - 2022-06-26